fix: refactor config/setup into own file
